pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Model of mislocalization and remapping"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.43",
    "tqdm>=4.66.5",
    "scipy>=1.14.1",
    "cmcrameri>=1.9",
    "ipympl>=0.9.6",
    "jupyterlab>=4.3.4",
    "equinox>=0.11.11",
    "diffrax>=0.6.2",
    "jupytext>=1.16.6",
    "matplotlib>=3.10.0",
    "statsmodels>=0.14.4",
]
authors = [
  {name = "Yohaï-Eliel Berreby", email = "me@yberreby.com"},
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
